+++ system/Pixel0.1/Pixel0.1/removeDuplicatePatchs.cpp
@@ -52,7 +52,7 @@ vector<SuperImagePatch*> removeDuplicateImagePatchs(vector<ImagePatch* >& patch_
     
     double nearest_score = 0.0;
     int nearest_index = -1;
-    for (int i = 1; i < patch_vec.size(); i++) {
+    for (size_t i = 1; i < patch_vec.size(); i++) {
         ImagePatch *one_patch = patch_vec[i];
         //convert first ,use second
         vector<Patch *> base_patch_vec = convert_verctor<Patch,SuperImagePatch>(supers);
@@ -79,10 +79,10 @@ vector<SuperImagePatch*> removeDuplicateImagePatchs(vector<ImagePatch* >& patch_
 vector<SuperImagePatch*> removeDuplicateImagePatch1To1(vector<ImagePatch* >& patch_vec){
     vector<SuperImagePatch *> result;
     bool find_similar ;
-    for (int i = 0 ; i < patch_vec.size(); i++) {
+    for (size_t i = 0 ; i < patch_vec.size(); i++) {
         find_similar = false;
         ImagePatch *one_patch = patch_vec[i];
-        for (int j = 0; j < result.size(); j++) {
+        for (size_t j = 0; j < result.size(); j++) {
             SuperImagePatch *tmp_sp = result[j];
             double score = one_patch->patchCompareWith(tmp_sp, Params::SHAPE_CONTEXT);
             // (1) has similar 'SP', insert
@@ -147,7 +147,7 @@ vector<SuperImagePatch*> removeDuplicateSuperImagePatchs(vector<SuperImagePatch*
     //compare each one in sp_vec to final_supers_vector
     double nearest_score = 0.0;
     int nearest_index = -1;
-    for (int i = 1; i < sp_vec.size(); i++) {
+    for (size_t i = 1; i < sp_vec.size(); i++) {
         SuperImagePatch *candi_sp = sp_vec[i];
         
         vector<Patch *> base = convert_verctor<Patch,SuperImagePatch>(final_sps);
